Create McDonald's Menu Analysis
This project uses SQLAlchemy, ibm_db, ibm_db_sa, seaborn, pandas and matplotlib in a python program made to analyze data about the food components of a McDonald's menu (database taken from Kaggle: https://www.kaggle.com/mcdonalds/nutrition-facts) and represent the findings in a graphical manner. The database is in the form of a csv file loaded onto the IBM DB2 cloud.
The graphs used are a scatter plot to visualize which the amount of particular food component present in a food item and a joint plot to see correlations between any two chosen food components.